+++ src/Saunter/AsyncApiMiddleware.cs
@@ -2,12 +2,10 @@
 
 using System;
 using System.Net;
-using System.Net.Mime;
-using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Options;
-using Saunter.Utils;
+using Saunter.Serialization;
 
 namespace Saunter
 {