+++ src/Saunter/AsyncApiServiceCollectionExtensions.cs
@@ -2,6 +2,7 @@ using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Saunter.Generation;
+using Saunter.Serialization;
 using Saunter.Utils;
 
 namespace Saunter