+++ src/Saunter/Generation/Filters/IOperationFilter.cs
@@ -14,16 +14,20 @@ namespace Saunter.Generation.Filters
 
     public class OperationFilterContext
     {
-        public OperationFilterContext(MethodInfo method, JsonSchemaResolver schemaResolver, OperationAttribute operation)
+        public OperationFilterContext(MethodInfo method, JsonSchemaResolver schemaResolver, JsonSchemaGenerator schemaGenerator, OperationAttribute operation)
         {
             Method = method;
             SchemaResolver = schemaResolver;
+            SchemaGenerator = schemaGenerator;
             Operation = operation;
         }
         
         public MethodInfo Method { get; }
+
         public JsonSchemaResolver SchemaResolver { get; }
-        
+
+        public JsonSchemaGenerator SchemaGenerator { get; }
+
         public OperationAttribute Operation { get; }
     }
 }