+++ test/Saunter.Tests/Utils/NJsonSchemaExtensions.cs
@@ -1,9 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using NJsonSchema;
 
-namespace Saunter.Utils
+namespace Saunter.Tests.Utils
 {
     public static class NJsonSchemaExtensions
     {