+++ CHANGELOG.md
@@ -6,9 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 <!-- Please update the links section at the bottom when adding a new version. -->
 
 ## [Unreleased]
+
+
+## [v0.5.0] - 2021-07-28
 ### Changed
 - Added ability to use bindings by specifying a BindingsRef in the attribute e.g. `[Channel("light.measured", BindingsRef = "my-amqp-binding")]`
 - Added MQTT bindings
+- Swapped custom JSON schema generation for NJsonSchema
+- Removed direct dependencies on System.Text.Json
+
+
 
 ## [v0.4.0] - 2021-07-27
 ### Changed